requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
click==8.1.4
colorama==0.4.6
fastapi==0.100.0
fastapi-csrf-protect==0.3.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.6
pydantic==1.10.12
pydantic_core==2.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
